+++ packages/design-system/CHANGELOG.md
@@ -12,7 +12,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
   - Add `size` prop to control the size of the button (e.g., 'sm', 'lg').
 - **DropdownButtonItem:**
   - Add `type` prop to allow specifying the type of the element.
-- **SelectAdvanced:** Fix word wrapping in options list to prevent overflow and ensure long text is displayed correctly.
+- **SelectAdvanced:**
+  - Fix word wrapping in options list to prevent overflow and ensure long text is displayed correctly.
+  - Support for options with a shape of `{ label: string; value: string; }[]` instead of just `string[]`.
 - **ButtonGroup:**
   - Fix styles for vertical button groups when using tooltips.